testlapack: move local implementations of Lapack functions to separate file
